requirement.txt
langchain
huggingface_hub
python-dotenv